[Refactor] isValid
The names of the methods weren't following the proper convention; test method names should be written like a continuation of the class name (which isValid isn't doing), therefore 'validate_all_password_requirements' is a better option. In the case of 'Validate', since its return type is a bool, normally it would names like a question (is<Property>).